components/devopsrunner/python/src/requirements.txt
termcolor==1.1.0
colorama==0.4.4
pyyaml==5.4.1
boto3==1.15.0
boto3_type_annotations==0.3.1
GitPython==3.1.14
semver>=2,<3
python-gitlab==2.6.0
wheel==0.37.1